Headless UI mode
This is an experimental new UI build target that is basically unix
console, but with all display operations replaced with noops. The idea
is to provide an alternative to `fake_pty` calls that can be a bit more
flexible about things like output, for example actually using stdout
(though this is not yet implemented). It's potentially also a step
towards the hypothetical libcrawl. When run from the command line it
requires some argument that will do something that makes sense in
headless, and if it receives a non-`-`'d first argument, will treat this
as a script. To build, pass HEADLESS=y to make. This is not compatible
with TILES=y or WEBTILES=y.

This commit is a proof-of-concept that compiles and works for at least
some scripts, including the seed explorer. It is also designed so that
switching between a headless and non-headless build is a matter of
swapping .o files, rather than rebuilding everything.  As of this
commit, this passes all the regular tests but won't run the stress tests
(because right now it doesn't know that they will work headlessly).

For webtiles builds, it may be possible to get this to work but it would
need more tweaking, as webtiles assumes that there is a tty for
recording purposes. I'm not sure if this is worth the effort.